CircleCI: Generate types for eslint
I stumbled upon this issue here when searching for what to do when PageData
types are not created:

https://github.com/sveltejs/kit/issues/5940

The solution is to run svelte-kit sync. I have added a small helper to our
package.json and I call it from the CircleCI check:eslint step.

Also including this in our bisect helper