Visualize bundle size in staging vite mode
We need to estimate the bundle size when optional packages are not
installed. That is why we have to run visualize-bundle (the name of the
CI step) using the production NODE_ENV.

One example for an optional package that will accidentally make it into
our bundle us moment.js through pikaday.

It would be great if package.json allowed you to specify depenency
groups.

On the other hand, the rollup bundle visualizer will only ever be
imported by vite when its mode is set to staging using an asynchronous
import() expression in vite.config.ts