Lib/Stores: Clear up context menu stack traces
If we return promises, instead of resolving them in each function we
lose a lot of useful debug information. This uncovered another issue in
turn: Context menus wait internally for the overlay to finish. That
shouldn't be the case, as a context menu state never should depend on
the result of whatever a user does with the overlay. We should close
the context menu immediately and hand off the rest to the other overlay.

See also the long stack trace in context-menu/WorkspaceBoard.svelte